+++ lib/f_service/result/base.rb
@@ -71,8 +71,8 @@ module FService
       # @yieldparam [Success] value the success value
       # @return [Success, Failure] the original Result object
       # @api public
-      def on_success
-        yield(value) if successful?
+      def on_success(target_type = nil)
+        yield(value) if successful? && expected_type?(target_type)
 
         self
       end
@@ -97,10 +97,17 @@ module FService
       # @yieldparam [Failure] failure the failure value
       # @return [Success, Failure] the original Result object
       # @api public
-      def on_failure
-        yield(error) if failed?
+      def on_failure(target_type = nil)
+        yield(error) if failed? && expected_type?(target_type)
 
         self
+      end
+
+      private
+
+      # @api private
+      def expected_type?(target_type)
+        type == target_type || target_type.nil?
       end
     end
   end